check enrollment count while enroll

+++ src/open-hackathon-server/Kaiyuanshe.OpenHackathon.ServerTests/Controllers/AwardControllerTests.cs
@@ -356,15 +356,17 @@ namespace Kaiyuanshe.OpenHackathon.ServerTests.Controllers
         #endregion
 
         #region DeleteAward
-        [Test]
-        public async Task DeleteAward_FirstTime()
+        [TestCase(true)]
+        [TestCase(false)]
+        public async Task DeleteAward(bool firstTimeDelete)
         {
             // input
             string hackName = "Foo";
             HackathonEntity hackathon = new HackathonEntity { };
             string awardId = "aid";
-            AwardEntity awardEntity = new AwardEntity { };
+            AwardEntity awardEntity = firstTimeDelete ? new AwardEntity() : null;
             CancellationToken cancellationToken = CancellationToken.None;
+            var authResult = AuthorizationResult.Success();
 
             // moq
             var hackathonManagement = new Mock<IHackathonManagement>();
@@ -373,53 +375,28 @@ namespace Kaiyuanshe.OpenHackathon.ServerTests.Controllers
             var awardManagement = new Mock<IAwardManagement>();
             awardManagement.Setup(t => t.GetAwardByIdAsync("foo", awardId, cancellationToken))
                 .ReturnsAsync(awardEntity);
-            awardManagement.Setup(t => t.DeleteAwardAsync(awardEntity, cancellationToken));
+            if (firstTimeDelete)
+            {
+                awardManagement.Setup(t => t.DeleteAwardAsync(awardEntity, cancellationToken));
+            }
+            var authorizationService = new Mock<IAuthorizationService>();
+            authorizationService.Setup(m => m.AuthorizeAsync(It.IsAny<ClaimsPrincipal>(), hackathon, AuthConstant.Policy.HackathonAdministrator))
+                .ReturnsAsync(authResult);
 
             // run
             var controller = new AwardController
             {
                 HackathonManagement = hackathonManagement.Object,
                 AwardManagement = awardManagement.Object,
+                AuthorizationService = authorizationService.Object,
             };
             var result = await controller.DeleteAward(hackName, awardId, cancellationToken);
 
             // verify
-            Mock.VerifyAll(hackathonManagement, awardManagement);
+            Mock.VerifyAll(hackathonManagement, awardManagement, authorizationService);
             hackathonManagement.VerifyNoOtherCalls();
             awardManagement.VerifyNoOtherCalls();
-            AssertHelper.AssertNoContentResult(result);
-        }
-
-        [Test]
-        public async Task DeleteAward_SecondTime()
-        {
-            // input
-            string hackName = "Foo";
-            HackathonEntity hackathon = new HackathonEntity { };
-            string awardId = "aid";
-            AwardEntity awardEntity = null;
-            CancellationToken cancellationToken = CancellationToken.None;
-
-            // moq
-            var hackathonManagement = new Mock<IHackathonManagement>();
-            hackathonManagement.Setup(p => p.GetHackathonEntityByNameAsync("foo", cancellationToken))
-                .ReturnsAsync(hackathon);
-            var awardManagement = new Mock<IAwardManagement>();
-            awardManagement.Setup(t => t.GetAwardByIdAsync("foo", awardId, cancellationToken))
-                .ReturnsAsync(awardEntity);
-
-            // run
-            var controller = new AwardController
-            {
-                HackathonManagement = hackathonManagement.Object,
-                AwardManagement = awardManagement.Object,
-            };
-            var result = await controller.DeleteAward(hackName, awardId, cancellationToken);
-
-            // verify
-            Mock.VerifyAll(hackathonManagement, awardManagement);
-            hackathonManagement.VerifyNoOtherCalls();
-            awardManagement.VerifyNoOtherCalls();
+            authorizationService.VerifyNoOtherCalls();
             AssertHelper.AssertNoContentResult(result);
         }
         #endregion
